% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{add_logo}
\alias{add_logo}
\title{Add A Logo}
\usage{
add_logo(ggplot = NULL, logo_path = NULL)
}
\arguments{
\item{ggplot}{name of the ggplot object you want to have a logo}

\item{logo_path}{file path to the logo. Must be a PNG file.}
}
\value{
a patchwork object
}
\description{
Add A Logo
}
\examples{
\dontrun{
library(mrworthingtonR)
library(ggplot2)

my_chart <- ggplot(mpg) +
  aes(x = displ, y = hwy, colour = class) +
  geom_point() +
  theme_mrw(legend_position = "none") +
  labs(title = "This is the title",
       subtitle = "This is a compelling subtitle",
       caption = "Where We Got The Data From")
logo_chart <- add_logo(my_chart)

logo_chart
}
}
