% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\docType{data}
\name{font_ft}
\alias{font_ft}
\title{Figtree font name R variable aliases}
\format{
length 1 character vector
}
\usage{
font_ft
}
\description{
\code{font_ft} == "\code{Figtree}"
}
\keyword{datasets}
