% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_basic_allotment.R
\name{get_ba}
\alias{get_ba}
\title{Get Historic Basic Allotment Data}
\source{
{\href{https://tea.texas.gov/sites/default/files/basic-allotment-one-pager.pdf}{Basic Allotment Explainer, Texas Education Agency}. State Funding Division, (512) 463-9238 or sfinance@tea.texas.gov}
}
\usage{
get_ba()
}
\value{
a tbl_df object
}
\description{
The basic allotment is the legislatively mandated apportionment of funds from the general revenue funds that goes to each school district to provide a basic level of education for the district’s residents. This data is based on recorded data found in the following \href{https://tea.texas.gov/sites/default/files/basic-allotment-one-pager.pdf}{PDF provided by the Texas Education Agency}. It represents the annual basic allotment figures for each school year, starting in 2010.
}
\examples{
library(mrworthingtonR)
library(dplyr)
ba_df <- get_ba() |>
  glimpse()

}
