% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation.R
\name{get_inflation}
\alias{get_inflation}
\title{Get Inflation Data from the Bureau of Labor Statistics}
\usage{
get_inflation(index = "cpi")
}
\arguments{
\item{index}{The index you want to pull from the Bureau of Labor Statistics. Options are "cpi" for All Consumers Price Index and "edu" for education + communication index. Default is CPI.}
}
\value{
a tbl_df object
}
\description{
Get Inflation Data from the Bureau of Labor Statistics
}
\examples{

library(mrworthingtonR)
library(dplyr)

cpi_raw <- get_inflation("cpi") |>
  filter(series_id == "CUSR0000SA0") |>
  select(year, period, cpi_value=value)

glimpse(cpi_raw)

edu_raw <- get_inflation("edu") |>
  filter(series_id == "CUSR0000SAE1") |>
  select(year, period, ed_value=value)

glimpse(edu_raw)

}
