% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{import_figtree}
\alias{import_figtree}
\title{Import Figtree font for use with ggplot2 Charts}
\usage{
import_figtree()
}
\description{
Figtree is \href{https://github.com/erikdkennedy/figtree/blob/master/OFL.txt}{licensed} under the Open Font License. You can use them in your products & projects – print or digital, commercial or otherwise. This isn't legal advice, please consider consulting a lawyer and see the full license for all details.
}
\details{
There is an option \code{mrworthington.loadfonts} which -- if set to \code{TRUE} -- will
call \code{extrafont::loadfonts()} to register non-core fonts with R PDF & PostScript
devices. If you are running under Windows, the package calls the same function
to register non-core fonts with the Windows graphics device.
}
\note{
This will take care of ensuring PDF/PostScript usage. The location of the
font directory is displayed after the base import is complete. It is highly
recommended that you install them on your system the same way you would any
other font you wish to use in other programs.
}
