% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{theme_mrw}
\alias{theme_mrw}
\title{Use ggplot2 Theme}
\usage{
theme_mrw(
  base_size = 14,
  base_family = "",
  title_size = 23,
  subtitle_size = 12,
  caption_size = 10,
  legend_position = "none",
  plot_title_position = "plot",
  ...
)
}
\arguments{
\item{base_size}{the base size of text for the plot}

\item{base_family}{the base font used generally across plot text}

\item{title_size}{the text size for the plot title}

\item{subtitle_size}{the text size for the plot subtitle}

\item{caption_size}{the text size for the plot caption}

\item{legend_position}{where the legend should be placed}

\item{plot_title_position}{position of plot title (either "plot" or "panel")}

\item{...}{Additional arguments passed to \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}}}
}
\value{
a ggplot theme
}
\description{
ggplot2 theme for use by me.
}
\note{
If planning to use "Figtree", you should download it from Google Fonts and use locally or use the showtext package.
}
\examples{
library(mrworthingtonR)
library(ggplot2)

ggplot(mpg) +
  aes(x = displ, y = hwy, colour = class) +
  geom_point() +
  theme_mrw(legend_position = "bottom") +
  labs(title = "This is the title",
       subtitle = "This is a compelling subtitle",
       caption = "Where We Got The Data From")
}
