% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{use_custom_fonts}
\alias{use_custom_fonts}
\title{Use LP Brand Fonts}
\usage{
use_custom_fonts(enable = TRUE)
}
\arguments{
\item{enable}{enable use of fonts}
}
\description{
Use LP Brand Fonts
}
\examples{
\dontrun{
library(mrworthingtonR)
library(ggplot2)
use_lp_fonts()

ggplot(mpg) +
  aes(x = displ, y = hwy, colour = class) +
  geom_point() +
  theme_mrw(legend_position = "bottom", base_family = "Figtree") +
  labs(title = "This is the title",
       subtitle = "This is a compelling subtitle",
       caption = "Where We Got The Data From")
}
}
